Retargeted C# project to .NET 4.0
